Clearer Organization
Still a lot to do to bring interfaces coming next
