+++ crates/systems/ai/src/memory/mod.rs
@@ -1,1 +1,8 @@
 pub mod types;
+
+/// Prelude for the memory module.
+///
+/// This includes types for storing and managing entity memories.
+pub mod prelude {
+    pub use crate::memory::types::{MemoryEvent, MemoryEventType, MemoryTimestamp};
+}
