+++ crates/systems/ai/src/personality/mod.rs
@@ -1,1 +1,8 @@
 pub mod traits;
+
+/// Prelude for the personality module.
+///
+/// This includes personality traits and related components.
+pub mod prelude {
+    pub use crate::personality::traits::Personality;
+}
